refacto dataValidator + dynamic cache + jwt

+++ app/controllers/subjectController.js
@@ -1,15 +1,16 @@
 const { Subject, User, Comment } = require("../models")
-const {deleteUnthorizedKeys, deleteEmptyKeys} = require("../services/utils")
-
-const validKeys = ["comments", "title", "description", "references"]
 
 module.exports = {
     getAllOrFilter: async (req, res) => {
         
         try {
-            let {author, title} = req.query
             
-            if(!author && !title){
+            
+            let {title} = req.query
+            
+            
+            
+            if(!title){
                 
                 const subjects = await Subject.find({})
                 .populate({path: "author", model: 'User', select: 'pseudo lastname firstname'})
@@ -23,10 +24,12 @@ module.exports = {
                 return res.json(subjects)
             }
             
-            if(!title) title = ""
-            if(title) title = title.toLowerCase()+title.toUpperCase()
             
-            const subjects = await Subject.find({$or: [{author}, {title: {$regex:`^[${title}]`}}]})
+           
+            
+            
+            
+            const subjects = await Subject.find({$or: [{title: {$regex:`${title}`, $options: 'i'} }]})
             .populate({path: "comments", model: 'Comment', select: '-subject', populate: 
                         {
                             path: 'author',
@@ -43,48 +46,23 @@ module.exports = {
         }
     },
 
-    getById: async (req, res) => {
-        const {_id} = req.params
-        try {
-
-            const subject = await Subject.findOne({_id})
-            .populate({path: "author", model: "User", select: 'pseudo lastname firstname'})
-            if(!subject) return res.status(404).json({message: "subject not found"})
-
-            res.json(subject)
-
-        } catch (error) { 
-            res.status(500).json(error)
-        }
-    },
-
-    getByAuthor: async (req, res) => {
-        const {author} = req.params
-
-        try {
-            
-            const subjects = await Subject.find({author})
-            .populate({path: "author", model: "User", select: 'pseudo lastname firstname'})
-            if(!subjects) return res.status(404).json({message: "subject not found"})
-
-            res.json(subjects)
-
-        } catch (error) {
-            res.status(500).json(error)
-        }
-        
-    },
 
     addOne: async (req, res) => {
         try {
-            const { author, title, description, date } = req.body
+            const { _id } = req.token
+         
     
-            const userExist = await User.exists({_id: author})
+            const userExist = await User.exists({ _id })
+            
             if(!userExist) return res.status(404).json({message: "l'utilisateur n'existe pas"})
 
             const newSubject = new Subject(req.body)
+
+            newSubject.author = _id
+           
             await newSubject.save()
-            await User.updateOne({ _id: author }, { $push: { subjects: newSubject._id } })
+
+            await User.updateOne({ _id }, { $push: { subjects: newSubject._id } })
             
             res.json(newSubject)
             
@@ -115,9 +93,12 @@ module.exports = {
 
     updateOne: async (req, res) => {
         try {
-            const {_id} = req.params
- 
-            const updatedSubject = await Subject.findOneAndUpdate({ _id }, req.body, {new: true})
+            const author = req.token._id
+            const {_id} = req.body
+           
+            
+            const updatedSubject = await Subject.findOneAndUpdate({ _id, author }, req.body, {new: true})
+           
             if(!updatedSubject) return res.status(404).json({message: "subject not found"})
             
             res.json(updatedSubject)
@@ -129,12 +110,13 @@ module.exports = {
 
     deleteOne: async (req, res) => {
         try {
-            const { _id } = req.params
-           
-            const subject = await Subject.findOneAndDelete({ _id })
+            const { _id } = req.query
+            const author = req.token._id
+            
+            const subject = await Subject.findOneAndDelete({ _id, author })
             if(!subject) return res.status(404).json({message: "subject not found"})
 
-            const {nModified} = await User.updateOne({ _id: subject.author }, { $pull: { subjects: _id } })
+            const {nModified} = await User.updateOne({ _id: author }, { $pull: { subjects: _id } })
 
             if(!nModified) return res.status(500).json({message: "author of subject not found"})
